mm (#7)
* Revert "Added DataUsageProvider to the build"

* CM specific packages should go into vendor/cm

This reverts commit 0c03d1e623e77bab4825da899e52b4ffc849a3af.

Change-Id: I5810111672a4566def56df80e9cdc9bd1e158c70

* repopick: support specifying a range of commits to pick

* for example: repopick 12345-12350

Change-Id: I3b8f2c331a6ac5488032432e133bd4b44d0bf007

* Avoid accidentally using the host's native 'as' command.

When invoking clang for the host to assemble .S files, the -B flag
would not be provided, which allowed the host prebuilt clang to
use an 'as' from the native environment. Most of the time this
"just works", but some newer 'as' versions cause problems with
the older prebuilt toolchain, for example by generating
unsupported relocation types.

To avoid this problem, simply use the -B flag to tell clang to
invoke the assembler from the correct prebuilt gcc prefix.

Change-Id: I18ea4f37ae637b652cfd7321c41929f8be075342
Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>

* releasetools: Allow for custom boot image signing tools

Some boot images will need to be signed using specific tools. Allow
passing the path to this tool through an environment variable.

Change-Id: I958726fdae8d63fe99d639dfa361046b528a087c
Ticket: CYNGNOS-2868