Mm6.0 (#10)
* Revert "Added DataUsageProvider to the build"

* CM specific packages should go into vendor/cm

This reverts commit 0c03d1e623e77bab4825da899e52b4ffc849a3af.

Change-Id: I5810111672a4566def56df80e9cdc9bd1e158c70

* repopick: support specifying a range of commits to pick

* for example: repopick 12345-12350

Change-Id: I3b8f2c331a6ac5488032432e133bd4b44d0bf007

* Avoid accidentally using the host's native 'as' command.

When invoking clang for the host to assemble .S files, the -B flag
would not be provided, which allowed the host prebuilt clang to
use an 'as' from the native environment. Most of the time this
"just works", but some newer 'as' versions cause problems with
the older prebuilt toolchain, for example by generating
unsupported relocation types.

To avoid this problem, simply use the -B flag to tell clang to
invoke the assembler from the correct prebuilt gcc prefix.

Change-Id: I18ea4f37ae637b652cfd7321c41929f8be075342
Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>

* releasetools: Allow for custom boot image signing tools

Some boot images will need to be signed using specific tools. Allow
passing the path to this tool through an environment variable.

Change-Id: I958726fdae8d63fe99d639dfa361046b528a087c
Ticket: CYNGNOS-2868

* qcom: Support 8937/8953

Change-Id: I84b1807e88d5be1230eab64bd1275a0677ce510f

* Add Bliss to ROM Installation

You  will need to add "ro.product.device" and  "ro.product.model" to your device tree.

Example
ro.build.product=shamu
ro.product.model=Nexus 6

placed inside "system.prop"

Change-Id: Ie25e318dde08ca623fe3e5775bda3085e538a8a2

* build: Fix support for missing line

Fixes build devices that don't have "model" line in build.prop

* Clean it up a bit

Change-Id: Ic783089342fc1d30996346cf12442ae44cc65cd1

* build: add kernel header dependency if module uses kernel headers

Many of the QCOM components use kernel headers, but don't declare
the dependency on them.  This is fine in CAF because of the way they
build the boot.img before anything else.  In CM, we don't build the
boot.img the same, so we run into a race between the kernel build &
these modules... and the modules lose.

Warn about modules that have this missing dependency, and add it for
them so we don't have to modify each Android.mk.

Change-Id: I95f1e47b5ef440f6f5d8f64a0c3f38d9572e839e